+++ src/reference.d.ts
@@ -18,6 +18,7 @@
 /// <reference path ="./interfaces/iServiceAction.ts" />
 /// <reference path ="./interfaces/iBoundBox.ts" />
 /// <reference path ="./interfaces/iVector.ts" />
+/// <reference path ="./interfaces/iEngineOptions.ts" />
 
 /// <reference path="./core/events/event.ts" />
 /// <reference path="./core/events/eventEmitter.ts" />